完成收货地址

+++ app/Http/Controllers/UserAddressesController.php
@@ -34,10 +34,12 @@ class UserAddressesController extends Controller
     }
     public function edit(UserAddress $user_address)
     {
+        $this->authorize('own', $user_address);
         return view('user_addresses.create_and_edit', ['address' => $user_address]);
     }
     public function update(UserAddress $user_address, UserAddressRequest $request)
     {
+        $this->authorize('own', $user_address);
         $user_address->update($request->only([
             'province',
             'city',
@@ -52,6 +54,7 @@ class UserAddressesController extends Controller
     }
     public function destroy(UserAddress $user_address)
     {
+        $this->authorize('own', $user_address);
         $user_address->delete();
 
         return [];
